requirements.txt
beautifulsoup4
django-extensions>=0.8
